# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API documentation",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/likes/{postId}/count": {
      "get": {
        "operationId": "countLikes",
        "tags": [
          "Like"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "count"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/likes/{postId}": {
      "post": {
        "operationId": "likePost",
        "tags": [
          "Like"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "number"
                    },
                    "postId": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "postId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "unlikePost",
        "tags": [
          "Like"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/courses/": {
      "post": {
        "operationId": "createCourse",
        "tags": [
          "Course"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "code",
                  "description"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listCourses",
        "tags": [
          "Course"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "code": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "code"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "operationId": "getCourseById",
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "operationId": "updateCourse",
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "operationId": "deleteCourse",
        "tags": [
          "Course"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333"
    }
  ]
}